Fix gh-1908: move Impl/ files from RabbitMQ.Client.Framing to RabbitMQ.Client.Impl namespace
Eight internal files in projects/RabbitMQ.Client/Impl/ declared namespace
RabbitMQ.Client.Framing despite living in the Impl/ directory:

- Connection.cs, Connection.Commands.cs, Connection.Heartbeat.cs, Connection.Receive.cs
- AutorecoveringConnection.cs, AutorecoveringConnection.Recovery.cs, AutorecoveringConnection.Recording.cs
- ProtocolBase.cs

Change all eight to namespace RabbitMQ.Client.Impl. Swap the existing
`using RabbitMQ.Client.Impl;` for `using RabbitMQ.Client.Framing;` in
the files that had it, and add the same using to Connection.Heartbeat.cs
which lacked it. Update Framing/Protocol.cs (which inherits ProtocolBase)
to import RabbitMQ.Client.Impl. Fix test files that referenced
AutorecoveringConnection via the old namespace.

No public API change - all affected types are internal.
