+++ projects/RabbitMQ.Client/Impl/Connection.Commands.cs
@@ -36,9 +36,9 @@ using System.Threading;
 using System.Threading.Tasks;
 using RabbitMQ.Client.Events;
 using RabbitMQ.Client.Exceptions;
-using RabbitMQ.Client.Impl;
+using RabbitMQ.Client.Framing;
 
-namespace RabbitMQ.Client.Framing
+namespace RabbitMQ.Client.Impl
 {
     internal sealed partial class Connection
     {
