+++ projects/Test/Integration/ConnectionRecovery/TestConsumerRecovery.cs
@@ -33,6 +33,7 @@ using System.Threading.Tasks;
 using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
 using RabbitMQ.Client.Framing;
+using RabbitMQ.Client.Impl;
 using Xunit;
 using Xunit.Abstractions;
