+++ projects/Test/Integration/ConnectionRecovery/TestRecoveryWithDeletedEntities.cs
@@ -34,6 +34,7 @@ using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
 using RabbitMQ.Client.Exceptions;
 using RabbitMQ.Client.Framing;
+using RabbitMQ.Client.Impl;
 using Xunit;
 using Xunit.Abstractions;
