+++ projects/Test/Integration/TestFloodPublishing.cs
@@ -65,7 +65,7 @@ namespace Test.Integration
             _connFactory.RequestedHeartbeat = TimeSpan.FromSeconds(60);
             _connFactory.AutomaticRecoveryEnabled = false;
             _conn = await _connFactory.CreateConnectionAsync();
-            Assert.IsNotType<RabbitMQ.Client.Framing.AutorecoveringConnection>(_conn);
+            Assert.IsNotType<RabbitMQ.Client.Impl.AutorecoveringConnection>(_conn);
             _channel = await _conn.CreateChannelAsync(_createChannelOptions);
 
             _conn.ConnectionShutdownAsync += (_, ea) =>
@@ -177,7 +177,7 @@ namespace Test.Integration
             _connFactory.AutomaticRecoveryEnabled = false;
 
             _conn = await _connFactory.CreateConnectionAsync();
-            Assert.IsNotType<RabbitMQ.Client.Framing.AutorecoveringConnection>(_conn);
+            Assert.IsNotType<RabbitMQ.Client.Impl.AutorecoveringConnection>(_conn);
             _channel = await _conn.CreateChannelAsync();
 
             string message = "Hello from test TestMultithreadFloodPublishing";
